Add texture size to configuration, implement game reset functionality, and refactor keyboard handling; update board size and apple counts in config
