+++ src/ui/gui.py
@@ -1,6 +1,6 @@
 """Game manager module."""
 
-from typing import List, Callable, Tuple, Optional, Dict
+from typing import Callable, Optional, Dict
 
 import pygame
 
@@ -9,13 +9,13 @@ from rich.console import Console
 
 from src.config.settings import Config
 from src.game.enviroment import Enviroment
-from src.game.snake import Snake
 from src.ui.texture_loader import TextureLoader
-from src.utils.keyboard_handler import KeyListener
 
 
-class Interpreter:
-    def __init__(self, config: Config, textures: TextureLoader) -> None:
+class GUI:
+    def __init__(
+        self, config: Config, textures: Optional[TextureLoader] = None
+    ) -> None:
         """
         Game manager class.
 
@@ -24,10 +24,14 @@ class Interpreter:
             textures (GameTextures): The game textures.
         """
         self.config: Config = config
-        self.textures: TextureLoader = textures
-        self.board: Enviroment = Enviroment(config=config)
-        self.game_visuals: str = self.config.visual.modes.mode
         self.game_controllers: Dict[int, Callable[[list[int]], None]] = {}
+        # <-- Visual Settings -->
+        self.textures: TextureLoader = textures or TextureLoader(config)
+        self.texture_size: int = config.pygame_textures.texture_size
+        self.game_visuals: str = self.config.visual.modes.mode
+        # <-- Game Data -->
+        self.board_width: int = self.config.map.board_size.width
+        self.board_height: int = self.config.map.board_size.height
         # <-- GUI Optionals -->
         self.console: Optional[Console] = None
         self.window: Optional[pygame.Surface] = None
@@ -36,82 +40,54 @@ class Interpreter:
         self.initialize()
 
     def initialize(self) -> None:
-        for snake in self.board.snakes:
-            self.game_controllers[snake.id] = snake.snake_controller
-        # self.board.add_apples()
-
         if self.game_visuals == "cli":
             self.console = Console()
             self.render: Callable = self.render_ascii
         else:
             pygame.init()
             self.window = pygame.display.set_mode(
-                size=(self.board.width * 32, self.board.height * 32)
+                size=(
+                    self.board_width * self.texture_size,
+                    self.board_height * self.texture_size,
+                )
             )
             self.render: Callable = self.render_pygame
 
-    def update(self) -> None:
-        self.board.add_apples()
-        for snake in self.board.snakes:
-            self.board.move_snake(snake=snake)
-            if self.board.check_collision(
-                snake=snake, snakes=self.board.snakes
-            ):
-                snake.alive = False
-            self.board.check_apple_eaten(snake)
-            self.board.update_snake_position(snake)
-
-        self.render()
-
-    def step(self, actions: List[list[int]], snakes: List[Snake]):
-        """
-        Perform one step in the game based on the chosen action,
-        then return the new state, reward, and done flag.
-
-        Args:
-            actions (List[list[int]]): Actions chosen by the agents
-                (0: up, 1: down, 2: left, 3: right).
-            snakes (List[Snake]): List of snake objects for the agents.
-
-        Returns:
-            Tuple[List[int], List[bool], List[int]]: The rewards, done flags, 
-                and sizes of all snakes.
-        """
-        self.board.step(actions)
-
-        self.render()
-
-        return self.board.rewards, self.board.snakes_done, self.board.snakes_done
-
-    def render_ascii(self) -> None:
+    def render_ascii(self, enviroment: Enviroment) -> None:
         if not self.console:
             print("Console not initialized.")
             return
         board_text = Text()
         self.console.clear()
-        for row in self.board.map:
+        for row in enviroment.map:
             for cell in row:
                 board_text += self.textures.textures[cell]
             board_text += "\n"
         self.console.print(board_text, justify="center")
-        for snake in self.board.snakes:
+        for snake in enviroment.snakes:
             self.console.print(
                 f"Snake {snake.id + 1}: "
                 f"Length: {len(snake.body)} | Kills: {snake.kills}",
                 justify="center",
             )
 
-    def render_pygame(self) -> None:
+    def render_pygame(self, enviroment: Enviroment) -> None:
         if not self.window:
             print("Window not initialized.")
             return
         self.window.fill(color=(18, 90, 60))
-        for row_num, row in enumerate(iterable=self.board.map):
+        for row_num, row in enumerate(iterable=enviroment.map):
             for col_num, cell in enumerate(iterable=row):
                 texture: pygame.Surface = pygame.image.load(
                     str(self.textures.textures[cell])
                 )
-                self.window.blit(texture, (col_num * 32, row_num * 32))
+                self.window.blit(
+                    texture,
+                    (
+                        col_num * self.texture_size,
+                        row_num * self.texture_size,
+                    ),
+                )
 
         score_positions = [
             (0, 0),
@@ -121,7 +97,7 @@ class Interpreter:
         ]
 
         font = pygame.font.Font(None, size=24)
-        for idx, snake in enumerate(iterable=self.board.snakes):
+        for idx, snake in enumerate(iterable=enviroment.snakes):
             if idx < len(score_positions):
                 text: pygame.Surface = font.render(
                     f"Snake {snake.id + 1}: "
@@ -132,11 +108,3 @@ class Interpreter:
                 self.window.blit(text, score_positions[idx])
 
         pygame.display.update()
-
-    def get_state_size(self) -> int:
-        return self.board.snake_state_size
-
-    def reset(self) -> None:
-        self.board = Enviroment(config=self.config)
-        for snake in self.board.snakes:
-            self.game_controllers[snake.id] = snake.snake_controller
